pyproject.toml
[project]
name = "telco-churn-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "flask>=3.1.2",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.1.0",
    "joblib>=1.5.3",
    "jupyter>=1.1.1",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.7",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "openai>=2.16.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "toml>=0.10.2",
    "uvicorn>=0.40.0",
]
